lightningd/plugin: Move structs to header, make 'paths_match' and 'plugin_kill' non static
